Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -5,8 +5,8 @@ int main()
     int a;
     int b;
     int c;
-    scanf("%d %d %d",&a,&b,&c);
-    double d= (a+b+c)/3;
-    printf("Average: %.2lf",d);
+    scanf("%f %f %f",&a,&b,&c);
+    float d= (a+b+c)/3;
+    printf("Average: %.2f",d);
     return 0;
 }
